\name{odis}
\alias{odis}
\alias{lodis}
\encoding{latin1}

\title{Orthogonal distances from a PCA or PLS score space}

\description{

\code{odis} calculates orthogonal distances (OD = "\eqn{X}-residuals") from a PCA, PLS or PLSR model, i.e. the Euclidean distances between row observations of a data set and their projection on the score space (see e.g. Hubert et al. 2005, Van Branden & Hubert 2005, p. 66; Varmuza & Filzmoser, 2009, p. 79).

\code{lodis} does the same calculation for each local model (i.e. for each new observation to predict) generated by functions \code{\link{locw}},  \code{\link{lwplsr}}, etc. 

}

\usage{

odis(fm, Xr, Xu = NULL)

lodis(fm, Xr, Xu)

}

\arguments{

\item{fm}{For \code{odis}, output of functions \code{\link{pca}}, \code{\link{pls}} or \code{\link{plsr}}. For \code{lodis}, output of functions  \code{\link{locw}},  \code{\link{lwplsr}}, etc.}

\item{Xr}{The matrix or data frame of reference (= training) observations that was used for building the preliminary model \code{fm}.}

\item{Xu}{A \eqn{m x p} matrix or data frame of new (= test) observations (\eqn{Xu} is not used in the calculation of the median and MAD used for calculating the standardized distances; see Details).}

}

\details{

A "standardized" dissimilarity is also given, calculated as \eqn{abs(d - median(d)) / mad(d)} where \eqn{d} is the distance. A value higher than 2.5 or 3 may be considered as extreme.

}

\references{

M. Hubert, P. J. Rousseeuw, K. Vanden Branden (2005). ROBPCA: a new approach to robust principal components analysis. Technometrics, 47, 64-79.

K. Vanden Branden, M. Hubert (2005). Robuts classification in high dimension based on the SIMCA method. Chem. Lab. Int. Syst, 79, 10-21. 

K. Varmuza, P. Filzmoser (2009). Introduction to multivariate statistical analysis in chemometrics. CRC Press, Boca Raton.

}

\examples{

n <- 8
p <- 6
set.seed(1)
X <- matrix(rnorm(n * p, mean = 10), ncol = p, byrow = TRUE)
y1 <- 100 * rnorm(n)
y2 <- 100 * rnorm(n)
Y <- cbind(y1, y2)
set.seed(NULL)

Xr <- X[1:6, ] ; Yr <- Y[1:6, ]
Xu <- X[7:8, ] ; Yu <- Y[7:8, ]

fm <- pls(Xr, Yr, ncomp = 3)
#fm <- plsr(Xr, Yr, Xu, ncomp = 3)

odis(fm, Xr)

odis(fm, Xr, Xu)

}

\keyword{datagen}