\name{odis}
\alias{odis}
\alias{lodis}
\encoding{latin1}

\title{Orthogonal distances from a PCA or PLS score space}

\description{

\code{odis} calculates orthogonal distances (OD = "\eqn{X}-residuals") from a PCA or PLS model, i.e. the Euclidean distances between the row observations of a data set and their projections on the score space (see e.g. Hubert et al. 2005, Van Branden & Hubert 2005, p. 66; Varmuza & Filzmoser, 2009, p. 79).

\code{lodis} does the same calculation for each local model (i.e. for each new observation to predict) generated by functions \code{\link{locw}},  \code{\link{lwplsr}}, etc. 

}

\usage{

odis(fm, Xr, Xu = NULL, typcut = c("param", "mad", "boxplot"))

lodis(fm, Xr, Xu, typcut = c("param", "mad", "boxplot"))

}

\arguments{

\item{fm}{For \code{odis}, output of functions \code{\link{pca}}, \code{\link{pls}} or \code{\link{plsr}}. For \code{lodis}, output of functions  \code{\link{locw}},  \code{\link{lwplsr}}, etc.}

\item{Xr}{The matrix or data frame of reference (= training) observations that was used for building the preliminary model \code{fm}.}

\item{Xu}{A \eqn{m x p} matrix or data frame of new (= test) observations (\eqn{Xu} is not used in the calculation of the median and MAD used for calculating the standardized distances; see Details).}

\item{typcut}{Character string defining the type of cutoff for the distance (see the code). Possible values are \code{"param"} (default) (parametric Gaussian quantile for the transformed distance, such as in Hubert et al. 2005 p.66, except that MCD location and scale estimates have been replaced by median and MAD), \code{"mad"} (\eqn{median(d) + 2.5 * mad(d)}) or \code{"boxplot"} (\eqn{Q3 + 1.5 IQR).}}

}

\details{

Column \code{dstand} in the output is a "standardized" OD defined as \eqn{OD / cutoff} where the \eqn{cutoff} is calculated such as in Hubert et al. (2005 p.66) (). A value \code{dstand > 1} may be considered as extreme.

}

\references{

M. Hubert, P. J. Rousseeuw, K. Vanden Branden (2005). ROBPCA: a new approach to robust principal components analysis. Technometrics, 47, 64-79.

K. Vanden Branden, M. Hubert (2005). Robuts classification in high dimension based on the SIMCA method. Chem. Lab. Int. Syst, 79, 10-21. 

K. Varmuza, P. Filzmoser (2009). Introduction to multivariate statistical analysis in chemometrics. CRC Press, Boca Raton.

}

\examples{

n <- 8
p <- 6
set.seed(1)
X <- matrix(rnorm(n * p, mean = 10), ncol = p, byrow = TRUE)
y1 <- 100 * rnorm(n)
y2 <- 100 * rnorm(n)
Y <- cbind(y1, y2)
set.seed(NULL)

Xr <- X[1:6, ] ; Yr <- Y[1:6, ]
Xu <- X[7:8, ] ; Yu <- Y[7:8, ]

fm <- pls(Xr, Yr, ncomp = 3)
#fm <- plsr(Xr, Yr, Xu, ncomp = 3)

odis(fm, Xr)

odis(fm, Xr, Xu)

}

\keyword{datagen}