\name{scordis}
\alias{scordis}
\alias{lscordis}
\encoding{latin1}

\title{Score distances (SD) in a PCA or PLS score space}

\description{

\code{scordis} calculates score distances (SD) from a PCA, PLS or PLSR model, i.e. the Mahalanobis distances between the projections of the row observations on the score space and the center of the score space.

\code{lscordis} does the same calculation for each local model (i.e. for each new observation to predict) generated by functions \code{\link{locw}},  \code{\link{lwplsr}}, etc. 

}

\usage{

scordis(fm)

lscordis(fm)

}

\arguments{

\item{fm}{For \code{scordis}, output of functions \code{\link{pca}}, \code{\link{pls}} or \code{\link{plsr}}. For \code{lscordis}, output of functions  \code{\link{locw}},  \code{\link{lwplsr}}, etc.}

}

\details{

A "standardized" dissimilarity is also given, calculated as \eqn{abs(d - median(d)) / mad(d)} where \eqn{d} is the distance. A value higher than 2.5 or 3 may be considered as extreme.

The Winisi GH is also provided ("extreme" if GH > 3).

}


\examples{

n <- 8
p <- 6
set.seed(1)
X <- matrix(rnorm(n * p, mean = 10), ncol = p, byrow = TRUE)
y1 <- 100 * rnorm(n)
y2 <- 100 * rnorm(n)
Y <- cbind(y1, y2)
set.seed(NULL)

Xr <- X[1:6, ] ; Yr <- Y[1:6, ]
Xu <- X[7:8, ] ; Yu <- Y[7:8, ]

fm <- pca(Xr, ncomp = 3)
#fm <- pls(Xr, Yr, ncomp = 3)
scordis(fm)

fm <- pca(Xr, Xu, ncomp = 3)
#fm <- pls(Xr, Yr, Xu, ncomp = 3)
#fm <- plsr(Xr, Yr, Xu, ncomp = 3)
scordis(fm)

}

\keyword{datagen}